+++ lucene/analysis/common/src/test/org/apache/lucene/analysis/synonym/word2vec/TestWord2VecSynonymProvider.java
@@ -21,7 +21,7 @@ import java.io.IOException;
 import java.util.List;
 import org.apache.lucene.tests.util.LuceneTestCase;
 import org.apache.lucene.util.BytesRef;
-import org.apache.lucene.util.TermAndBoost;
+import org.apache.lucene.util.QueryBuilder.TermAndBoost;
 import org.apache.lucene.util.TermAndVector;
 import org.junit.Test;
